blog: Do not display tags that start with a dot
To be able to better categorise which post goes on which planet, I use
tags. But I do not neccessarily want to display those tags publicly, so
with this patch, all tags that start with a dot, will be hidden.

madness will generate an index for them and a feed, but posts will not
show them, neither will they appear in the global taglist in the footer.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>